reuse variation click logic
